+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 = 1.8.0 =
 
 Added:
+* Support for RSS feeds
 * Auto crop option. If an image width matches the image height - auto crop the image.
 * Use img width size. New option that allows using the img width attribute value for the image size, if the value is smaller than the requested image.
 * Bulk remove files from the media library.